app/requirements.txt
boto3
flask
faiss-cpu
numpy
nltk
sentence-transformers
google-genai
torch==2.5.1 --extra-index-url https://download.pytorch.org/whl/cpu